refactor: useGetQuestions 훅 리펙토링
