+++ src/libs/hooks/useGetQuestions.js
@@ -1,0 +1,16 @@
+import supabase from '@/libs/api/supabase.api';
+import { useState, useEffect } from 'react';
+
+export default function useGetQuestions() {
+  const [questions, setQuestions] = useState([]);
+
+  useEffect(() => {
+    async function fetchQuestions() {
+      const { data, error } = await supabase.from('password_questions').select();
+      if (!error) setQuestions(data);
+    }
+
+    fetchQuestions();
+  }, []);
+  return { questions };
+}
